+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/controller/ConfiguredApiController.kt
@@ -14,6 +14,7 @@ import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.Authentication
 import org.springframework.validation.annotation.Validated
+import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestParam
@@ -22,6 +23,7 @@ import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.Configu
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.FILTERS_QUERY_DESCRIPTION
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.ConfiguredApiController.FiltersPrefix.FILTERS_QUERY_EXAMPLE
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.controller.model.Count
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementCancellationResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
 import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.exception.NoDataAvailableException
@@ -194,16 +196,6 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
       "Athena automatically retries your queries in cases of certain transient errors. " +
       "As a result, you may see the query state transition from STARTED or FAILED to SUBMITTED.\n",
     security = [ SecurityRequirement(name = "bearer-jwt") ],
-    responses = [
-      ApiResponse(
-        headers = [
-          Header(
-            name = NO_DATA_WARNING_HEADER_NAME,
-            description = "Provides additional information about why no data has been returned.",
-          ),
-        ],
-      ),
-    ],
   )
   fun getQueryExecutionStatus(
     @PathVariable("reportId") reportId: String,
@@ -217,21 +209,35 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     dataProductDefinitionsPath: String? = null,
     authentication: Authentication,
   ): ResponseEntity<StatementExecutionStatus> {
-    return try {
-      ResponseEntity
-        .status(HttpStatus.OK)
-        .body(
-          configuredApiService.getStatementStatus(statementId, reportId, reportVariantId, dataProductDefinitionsPath),
-        )
-    } catch (exception: NoDataAvailableException) {
-      val headers = HttpHeaders()
-      headers[NO_DATA_WARNING_HEADER_NAME] = singletonList(exception.reason)
+    return ResponseEntity
+      .status(HttpStatus.OK)
+      .body(
+        configuredApiService.getStatementStatus(statementId, reportId, reportVariantId, dataProductDefinitionsPath),
+      )
+  }
 
-      ResponseEntity
-        .status(HttpStatus.OK)
-        .headers(headers)
-        .body(null)
-    }
+  @DeleteMapping("/reports/{reportId}/{reportVariantId}/statements/{statementId}")
+  @Operation(
+    description = "Cancels the execution of a running query.",
+    security = [ SecurityRequirement(name = "bearer-jwt") ],
+  )
+  fun cancelQueryExecution(
+    @PathVariable("reportId") reportId: String,
+    @PathVariable("reportVariantId") reportVariantId: String,
+    @PathVariable("statementId") statementId: String,
+    @Parameter(
+      description = ReportDefinitionController.DATA_PRODUCT_DEFINITIONS_PATH_DESCRIPTION,
+      example = ReportDefinitionController.DATA_PRODUCT_DEFINITIONS_PATH_EXAMPLE,
+    )
+    @RequestParam("dataProductDefinitionsPath", defaultValue = ReportDefinitionController.DATA_PRODUCT_DEFINITIONS_PATH_EXAMPLE)
+    dataProductDefinitionsPath: String? = null,
+    authentication: Authentication,
+  ): ResponseEntity<StatementCancellationResponse> {
+    return ResponseEntity
+      .status(HttpStatus.OK)
+      .body(
+        configuredApiService.cancelStatementExecution(statementId, reportId, reportVariantId, dataProductDefinitionsPath),
+      )
   }
 
   @GetMapping("/report/tables/{tableId}/count")
@@ -275,16 +281,6 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     description = "Returns the resulting rows of the executed statement in a paginated " +
       "fashion which has been have been stored in a dedicated table.",
     security = [ SecurityRequirement(name = "bearer-jwt") ],
-    responses = [
-      ApiResponse(
-        headers = [
-          Header(
-            name = NO_DATA_WARNING_HEADER_NAME,
-            description = "Provides additional information about why no data has been returned.",
-          ),
-        ],
-      ),
-    ],
   )
   fun getQueryExecutionResult(
     @PathVariable("reportId") reportId: String,
@@ -300,28 +296,18 @@ class ConfiguredApiController(val configuredApiService: ConfiguredApiService) {
     pageSize: Long,
     authentication: Authentication,
   ): ResponseEntity<List<Map<String, Any?>>> {
-    return try {
-      ResponseEntity
-        .status(HttpStatus.OK)
-        .body(
-          configuredApiService.getStatementResult(
-            tableId,
-            reportId,
-            reportVariantId,
-            dataProductDefinitionsPath,
-            selectedPage,
-            pageSize,
-          ),
-        )
-    } catch (exception: NoDataAvailableException) {
-      val headers = HttpHeaders()
-      headers[NO_DATA_WARNING_HEADER_NAME] = singletonList(exception.reason)
-
-      ResponseEntity
-        .status(HttpStatus.OK)
-        .headers(headers)
-        .body(null)
-    }
+    return ResponseEntity
+      .status(HttpStatus.OK)
+      .body(
+        configuredApiService.getStatementResult(
+          tableId,
+          reportId,
+          reportVariantId,
+          dataProductDefinitionsPath,
+          selectedPage,
+          pageSize,
+        ),
+      )
   }
 
   @GetMapping("/reports/{reportId}/{reportVariantId}/{fieldId}")